// src/app/E-Training_API/models/formation.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { FormationFormateur } from '../models/formation-formateur';
import { Inscription } from '../models/inscription';
import { Utilisateur } from '../models/utilisateur';
export interface Formation {
  DateDebut?: string;
  DateFin?: string;
  Description?: string | null;
  Duree?: string | null;
  EstInscrit?: boolean;
  Formateurs?: Array<Utilisateur> | null;
  FormationFormateurs?: Array<FormationFormateur> | null;
  Id?: string | null;
  InscriptionId?: string | null;
  Inscriptions?: Array<Inscription> | null;
  PlacesDisponibles?: number;
  Salle: string;
  Titre: string;
}