// src/app/E-Training_API/models/create-formation-view-model.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreateFormationViewModel {
  DateDebut: string;
  DateFin: string;
  Description: string;
  FormateurIds?: Array<string> | null;
  Id?: string | null;
  PlacesDisponibles: number;
  Salle: string;
  Titre: string;
}